moved errors back to api to fix ES6 syntax build issues
