+++ packages/altair-api/src/common/errors.ts
